Moved around util files to util folder in ast2ram.
