+++ src/Makefile.am
@@ -200,17 +200,17 @@ souffle_sources = \
         ast2ram/ClauseTranslator.h                         \
         ast2ram/ConstraintTranslator.cpp                   \
         ast2ram/ConstraintTranslator.h                     \
-        ast2ram/Location.h                                 \
         ast2ram/ProvenanceClauseTranslator.cpp             \
         ast2ram/ProvenanceClauseTranslator.h               \
         ast2ram/ProvenanceTranslator.cpp                   \
         ast2ram/ProvenanceTranslator.h                     \
+        ast2ram/utility/Location.h                         \
         ast2ram/utility/Utils.cpp                          \
         ast2ram/utility/Utils.h                            \
         ast2ram/utility/TranslatorContext.cpp              \
         ast2ram/utility/TranslatorContext.h                \
-        ast2ram/ValueIndex.cpp                             \
-        ast2ram/ValueIndex.h                               \
+        ast2ram/utility/ValueIndex.cpp                     \
+        ast2ram/utility/ValueIndex.h                       \
         ast2ram/ValueTranslator.cpp                        \
         ast2ram/ValueTranslator.h                          \
         interpreter/Context.h                              \
