+++ src/ast2ram/ClauseTranslator.cpp
@@ -32,11 +32,11 @@
 #include "ast/utility/Visitor.h"
 #include "ast2ram/AstToRamTranslator.h"
 #include "ast2ram/ConstraintTranslator.h"
-#include "ast2ram/Location.h"
-#include "ast2ram/ValueIndex.h"
 #include "ast2ram/ValueTranslator.h"
+#include "ast2ram/utility/Location.h"
 #include "ast2ram/utility/TranslatorContext.h"
 #include "ast2ram/utility/Utils.h"
+#include "ast2ram/utility/ValueIndex.h"
 #include "ram/Aggregate.h"
 #include "ram/Break.h"
 #include "ram/Conjunction.h"
