+++ src/ast2ram/ConstraintTranslator.cpp
@@ -18,10 +18,10 @@
 #include "ast/TranslationUnit.h"
 #include "ast/analysis/AuxArity.h"
 #include "ast2ram/AstToRamTranslator.h"
-#include "ast2ram/ValueIndex.h"
 #include "ast2ram/ValueTranslator.h"
 #include "ast2ram/utility/TranslatorContext.h"
 #include "ast2ram/utility/Utils.h"
+#include "ast2ram/utility/ValueIndex.h"
 #include "ram/Constraint.h"
 #include "ram/EmptinessCheck.h"
 #include "ram/ExistenceCheck.h"
