+++ src/ast2ram/ProvenanceClauseTranslator.cpp
@@ -20,9 +20,9 @@
 #include "ast/Clause.h"
 #include "ast/ProvenanceNegation.h"
 #include "ast2ram/AstToRamTranslator.h"
-#include "ast2ram/ValueIndex.h"
 #include "ast2ram/ValueTranslator.h"
 #include "ast2ram/utility/TranslatorContext.h"
+#include "ast2ram/utility/ValueIndex.h"
 #include "ram/Condition.h"
 #include "ram/Relation.h"
 #include "ram/SignedConstant.h"
