+++ src/ast2ram/ValueTranslator.cpp
@@ -19,9 +19,9 @@
 #include "ast/StringConstant.h"
 #include "ast/UserDefinedFunctor.h"
 #include "ast/analysis/Functor.h"
-#include "ast2ram/ValueIndex.h"
 #include "ast2ram/utility/TranslatorContext.h"
 #include "ast2ram/utility/Utils.h"
+#include "ast2ram/utility/ValueIndex.h"
 #include "ram/AutoIncrement.h"
 #include "ram/FloatConstant.h"
 #include "ram/IntrinsicOperator.h"
